+++ src/zabbix_server/timer/timer.c
@@ -644,7 +644,7 @@ ZBX_THREAD_ENTRY(timer_thread, args)
 			update_time -= update_time % 60;
 			maintenance_time = update_time;
 
-			if (0 > (idle = ZBX_TIMER_DELAY - (zbx_time() - maintenance_time)))
+			if (0 > (idle = ZBX_TIMER_DELAY - (int)(zbx_time() - maintenance_time)))
 				idle = 0;
 
 			zbx_setproctitle("%s #%d [%s, idle %d sec]",
